Fix bus error on test NIfTI creation
When loading a NIfTI image, a mmap is created to its data. When
overwriting that same image using the mmap'd data, sometimes a bus error
can occur. This is because the file is truncated when writing starts, so
the mmap doesn't point to data anymore. Loading the file into memory
(mmap=False) fixes this issue.