Remove get and append incremental from public API
appendIncremental has quadratic time complexity in the size of the image
being appended to. Because the predominant use case of RT-Cloud is
real-time fMRI, in which full fMRI scan runs are done at one time, the
primary interface to access BidsArchive for performance should be
BidsRun. Thus, we remove these methods from the public API in order to
encourage use of BidsRun and make explicit that get/appendIncremental
are not supported.