Add entity updating to BidsRun
This commit implements the following behavior: When users create a
BidsRun, they should be able to specify a set of entities that determine
which BidsIncrementals can be appended to the run. When the
BidsIncrementals are actually appended, if the BidsIncremental matches
the run's entities but also carries additional ones (e.g., a run number
or datatype for a BidsRun that only has 'subject' and 'task' defined),
the BidsRun should update to include these new, non-conflicting
entities.