Optimize get single incremental case
When getting a single incremental, using the Nibabel ArrayProxy is more
efficient than using the Numpy array directly. This is because the
ArrayProxy only reads the relevant part on disk, whereas getting the
full Numpy array reads a substantial amount of un-needed data. While
the latter is more efficient for BidsRun, as the disk access time is
amortized across more data, here it's more efficient just to use the
ArrayProxy.