Switch to directly slicing dataobj to avoid unnecessary copy
Using getNiftiData, which utilized np.asanyarray to get the dataobj as a numpy
array resulted in unexpected behavior:
1) When the dataobj was a np.memmap, it would be passed through, and slicing
the array would result in no data being read in to memory
2) When the dataobj was not a memmap, a copy would be performed, even if the
data had already been read in to memory.

This change uses a lazy load on the getBidsRun side and does the actual memory
reading using direct slicing of the dataobj, ensuring all memory is only read
once.